correções na pasta domain
