--- neos-spring-java/src/main/java/school/sptech/neosspringjava/DataBase/testmock.txt
@@ -1,46 +1,0 @@
-
-[
-{
-	"name":"gabriel",
-	"email":"email",
-	"password":"password",
-	"permission":1,
-	"telephone":"telephone",
-	"cpf":"cpf"
-},
-{
-	"name":"gabriel",
-	"email":"email",
-	"password":"password",
-	"permission":1,
-	"telephone":"telephone",
-	"cpf":"cpf"
-},
-{
-	"name":"gabriel",
-	"email":"email",
-	"password":"password",
-	"permission":1,
-	"telephone":"telephone",
-	"cpf":"cpf"
-},
-{
-	"name":"gabriel",
-	"email":"email",
-	"password":"password",
-	"permission":1,
-	"telephone":"telephone",
-	"cpf":"cpf"
-}
-]
-
-
-
-localhost:8080/user
-
-localhost:8080/user/list-of-users
-
-localhost:8080/user/3
-{
-	"name":"user update nome "
-}
