+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/domain/couponAvailable/CouponAvailable.java
@@ -1,4 +1,4 @@
-package school.sptech.neosspringjava.entity;
+package school.sptech.neosspringjava.domain.couponAvailable;
 
 import java.sql.Date;   
 import jakarta.persistence.Entity;
