+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/domain/employee/Employee.java
@@ -1,4 +1,5 @@
-package school.sptech.neosspringjava.entity;
+
+package school.sptech.neosspringjava.domain.employee;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -11,7 +12,8 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-public class Client {
+public class Employee {
+
 
     @Id
     @Getter
@@ -34,13 +36,21 @@ public class Client {
     private String passaword;
     @Getter
     @Setter
-    private Integer fkLocal;
-    @Override
-    public String toString() {
-        return "Client [id=" + id + ", name=" + name + ", email=" + email + ", passaword=" + passaword + ", fkLocal="
-                + fkLocal + "]";
-    }
+    @NotBlank(message = "um funcionário precisa de um estabelecimento")
+    @NotEmpty(message = "um funcionário precisa de um estabelecimento")
+    private Integer fkEstablishment;
+    @Getter
+    @Setter
+    @NotBlank(message = "um funcionário precisa de um estabelecimento")
+    @NotEmpty(message = "um funcionário precisa de um estabelecimento")
+    private Integer fkEmployeeType;
 
+    // @Override
+    // public String toString() {
+    //     return "Employee [id=" + id + ", name=" + name + ", email=" + email + ", passaword=" + passaword
+    //             + ", fkEstablishment=" + fkEstablishment + ", fkEmployeeType=" + fkEmployeeType + "]";
+    // }
 
     
+
 }
