+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/domain/establishment/Establishment.java
@@ -1,4 +1,4 @@
-package school.sptech.neosspringjava.entity;
+package school.sptech.neosspringjava.domain.establishment;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -7,32 +7,30 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-@Table(name = "address")
-public class Address {
-
+@Table(name = "Establishment")
+public class Establishment {
     @Id
     @Getter
     @Setter
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idAddress;
+    private int idEstablishment;
+    @Getter
+    @Setter
+    @NotBlank(message = "Nome é obrigatório")
+    @NotEmpty(message = "Nome é obrigatório")
+    private String name;
 
     @Getter
     @Setter
-    @NotEmpty
-    @NotBlank(message = "logradouro é obrigatório")
-    private String street;
+    @NotNull(message = "FkEmpresa é obrigatório")
+    private int fkEstablishment;
     @Getter
     @Setter
-    @NotEmpty
-    @NotBlank(message = "cidade é obrigatório")
-    private String city;
-    @Getter
-    @Setter
-    @NotEmpty
-    @NotBlank(message = "estado é obrigatório")
-    private String state;
+    @NotNull(message = "FkLocal é obrigatório")
+    private int fkLocal;
 }
