+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/domain/filter/Filter.java
@@ -1,10 +1,9 @@
-package school.sptech.neosspringjava.entity;
+package school.sptech.neosspringjava.domain.filter;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.Table;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
@@ -12,25 +11,31 @@ import lombok.Getter;
 import lombok.Setter;
 
 @Entity
-@Table(name = "Establishment")
-public class Establishment {
+public class Filter {
     @Id
     @Getter
     @Setter
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idEstablishment;
+    private Integer id;
     @Getter
     @Setter
-    @NotBlank(message = "Nome é obrigatório")
-    @NotEmpty(message = "Nome é obrigatório")
-    private String name;
-
+    @NotNull
+    @NotBlank
+    @NotEmpty
+    private Double preco;
     @Getter
     @Setter
-    @NotNull(message = "FkEmpresa é obrigatório")
-    private int fkEstablishment;
+    private Integer fkEstablishment;
     @Getter
     @Setter
-    @NotNull(message = "FkLocal é obrigatório")
-    private int fkLocal;
+    @NotNull
+    @NotBlank
+    @NotEmpty
+    private Integer fkService;
+    @Override
+    public String toString() {
+        return "Filter [id=" + id + ", preco=" + preco + ", fkEstablishment=" + fkEstablishment + ", fkService="
+                + fkService + "]";
+    }
+    
 }
