+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/domain/service/Service.java
@@ -1,4 +1,4 @@
-package school.sptech.neosspringjava.entity;
+package school.sptech.neosspringjava.domain.service;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
