+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/AdressRepository.java
@@ -2,7 +2,7 @@ package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import school.sptech.neosspringjava.entity.Address;
+import school.sptech.neosspringjava.domain.address.Address;
 
 public interface AdressRepository extends JpaRepository<Address, Integer>{
