+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/ClientRepository.java
@@ -1,7 +1,8 @@
 package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import school.sptech.neosspringjava.entity.Client;
+
+import school.sptech.neosspringjava.domain.client.Client;
 
 public interface ClientRepository extends JpaRepository<Client, Integer> {
