+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/CompanyRepository.java
@@ -2,7 +2,7 @@ package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import school.sptech.neosspringjava.entity.Company;
+import school.sptech.neosspringjava.domain.company.Company;
 
 public interface CompanyRepository extends JpaRepository<Company, Integer>{
