+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/CouponAvaliableRepository.java
@@ -2,7 +2,7 @@ package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import school.sptech.neosspringjava.entity.CouponAvailable;
+import school.sptech.neosspringjava.domain.couponAvailable.CouponAvailable;
 
 public interface CouponAvaliableRepository extends JpaRepository<CouponAvailable,Integer> {
