+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/CouponRepository.java
@@ -2,7 +2,7 @@ package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import school.sptech.neosspringjava.entity.Coupon;
+import school.sptech.neosspringjava.domain.coupon.Coupon;
 
 public interface CouponRepository extends JpaRepository <Coupon , Integer>{
