+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/DiscountTypeRepository.java
@@ -2,7 +2,7 @@ package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import school.sptech.neosspringjava.entity.DiscountType;
+import school.sptech.neosspringjava.domain.discountType.DiscountType;
 
 public interface DiscountTypeRepository extends JpaRepository<DiscountType, Integer>{
