+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/EmployeeRepository.java
@@ -2,7 +2,7 @@ package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import school.sptech.neosspringjava.entity.Employee;
+import school.sptech.neosspringjava.domain.employee.Employee;
 
 public interface EmployeeRepository extends JpaRepository<Employee, Integer>{
