+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/EstablishmentRopository.java
@@ -1,7 +1,8 @@
 package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import school.sptech.neosspringjava.entity.Establishment;
+
+import school.sptech.neosspringjava.domain.establishment.Establishment;
 public interface EstablishmentRopository extends JpaRepository<Establishment, Integer>{
 
 }
