+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/FilterRepository.java
@@ -2,7 +2,7 @@ package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import school.sptech.neosspringjava.entity.Filter;
+import school.sptech.neosspringjava.domain.filter.Filter;
 
 
 public interface FilterRepository extends JpaRepository <Filter, Integer>{
