+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/LocalRepository.java
@@ -1,7 +1,8 @@
 package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import school.sptech.neosspringjava.entity.Local;
+
+import school.sptech.neosspringjava.domain.local.Local;
 
 public interface LocalRepository extends JpaRepository<Local, Integer>{
