+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/ProductTypeRopository.java
@@ -2,7 +2,7 @@ package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import school.sptech.neosspringjava.entity.ProductType;
+import school.sptech.neosspringjava.domain.productType.ProductType;
 
 public interface ProductTypeRopository extends JpaRepository <ProductType, Integer>{
