+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/ServiceRepository.java
@@ -2,7 +2,7 @@ package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import school.sptech.neosspringjava.entity.Service;
+import school.sptech.neosspringjava.domain.service.Service;
 
 public interface ServiceRepository extends JpaRepository<Service, Integer>{
