+++ neos-spring-java/src/main/java/school/sptech/neosspringjava/repository/ServiceTypeRepository.java
@@ -2,7 +2,7 @@ package school.sptech.neosspringjava.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import school.sptech.neosspringjava.entity.ServiceType;
+import school.sptech.neosspringjava.domain.serviceType.ServiceType;
 
 public interface ServiceTypeRepository extends JpaRepository<ServiceType,Integer > {
